requirements.txt
imageio==2.25.1
imageio-ffmpeg==0.4.8
numpy==1.24.2
Pillow==9.4.0
